rename ViewNoFragmentComponent.class to ViewComponent.class and ViewNoFragmentComponentBuilder.class to ViewComponentBuilder.class
RELNOTES=N/A

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=310473988
